move variable_container from process to sage_scan
