+++ sage_scan/process/variable_resolver.py
@@ -252,77 +252,6 @@ def flatten_vars_dict(vars_dict: dict, _prefix: str = ""):
     return flat_vars_dict
 
 
-def extract_variable_names(txt):
-    if not variable_block_re.search(txt):
-        return []
-    found_var_blocks = variable_block_re.findall(txt)
-    blocks = []
-    for b in found_var_blocks:
-        if "lookup(" in b.replace(" ", ""):
-            continue
-        parts = b.split("|")
-        var_name = ""
-        default_var_name = ""
-        for i, p in enumerate(parts):
-            if i == 0:
-                var_name = p.replace("{{", "").replace("}}", "")
-                if " if " in var_name and " else " in var_name:
-                    # this block is not just a variable, but an expression
-                    # we need to split this with a space to get its elements
-                    skip_elements = ["if", "else", "+", "is", "defined"]
-                    sub_parts = var_name.split(" ")
-                    for sp in sub_parts:
-                        if not sp:
-                            continue
-                        if sp and sp in skip_elements:
-                            continue
-                        if sp and sp[0] in ['"', "'"]:
-                            continue
-                        var_name = sp
-                        break
-                var_name = var_name.replace(" ", "")
-                if "lookup(" in var_name and "first_found" in var_name:
-                    var_name = var_name.split(",")[-1].replace(")", "")
-                if var_name and var_name[0] == "(":
-                    var_name = var_name.split(")")[0].replace("(", "")
-                if "+" in var_name:
-                    sub_parts = var_name.split("+")
-                    for sp in sub_parts:
-                        if not sp:
-                            continue
-                        if sp and sp[0] in ['"', "'"]:
-                            continue
-                        var_name = sp
-                        break
-                if "[" in var_name and "." not in var_name:
-                    # extract dict/list name
-                    dict_pattern = r'(\w+)\[(\'|").*(\'|")\]'
-                    match = re.search(dict_pattern, var_name)
-                    if match:
-                        matched_str = match.group(1)
-                        var_name = matched_str.split("[")[0]
-                    list_pattern = r'(\w+)\[\-?\d+\]'
-                    match = re.search(list_pattern, var_name)
-                    if match:
-                        matched_str = match.group(1)
-                        var_name = matched_str.split("[")[0]
-            else:
-                if "default(" in p and ")" in p:
-                    default_var = p.replace("}}", "").replace("default(", "").replace(")", "").replace(" ", "")
-                    if not default_var.startswith('"') and not default_var.startswith("'") and not re.compile(r"[0-9].*").match(default_var):
-                        default_var_name = default_var
-        tmp_b = {
-            "original": b,
-        }
-        if var_name == "":
-            continue
-        tmp_b["name"] = var_name
-        if default_var_name != "":
-            tmp_b["default"] = default_var_name
-        blocks.append(tmp_b)
-    return blocks
-
-
 def resolved_vars_contains(resolved_vars, new_var):
     if not isinstance(new_var, dict):
         return False
